added namespaces a daq datamodel base class  and changed makefile.
