+++ UserTools/ServiceAdd/ServiceAdd.h
@@ -10,6 +10,7 @@
 #include <boost/uuid/uuid_generators.hpp> // generators                                                                     
 #include <boost/uuid/uuid_io.hpp>         // streaming operators etc.        
 
+#include "DataModel.h"
 
 /**
  * \class ServiceAdd
