+++ src/DAQDataModelBase/DAQDataModelBase.cpp
@@ -1,0 +1,29 @@
+#include "DAQDataModelBase.h"
+
+using namespace ToolFramework;
+
+DAQDataModelBase::DAQDataModelBase(): DataModelBase() {}
+
+/*
+TTree* DAQDataModelBase::GetTTree(std::string name){
+
+  return m_trees[name];
+
+}
+
+
+void DAQDataModelBase::AddTTree(std::string name,TTree *tree){
+
+  m_trees[name]=tree;
+
+}
+
+
+void DAQDataModelBase::DeleteTTree(std::string name,TTree *tree){
+
+  m_trees.erase(name);
+
+}
+
+*/
+
