+++ src/DAQDataModelBase/SlowControlCollection.h
@@ -6,31 +6,33 @@
 #include "DAQUtilities.h"
 #include <functional>
 
-class SlowControlCollection;
-
-struct SlowControlCollectionThread_args:Thread_args{
-
+namespace ToolFramework{
+  
+  class SlowControlCollection;
+  
+  struct SlowControlCollectionThread_args:Thread_args{
+  
   SlowControlCollectionThread_args();
   ~SlowControlCollectionThread_args();
-
+  
   zmq::socket_t* sock;
   zmq::socket_t* sub;
   zmq::pollitem_t items[2];
   int poll_length;
-
+  
   SlowControlCollection* SCC;
   std::map<std::string, std::function<void(const char*)> >* trigger_functions;
   std::mutex* trigger_functions_mutex;
-
+  
 };
-
-class SlowControlCollection{
-
+  
+  class SlowControlCollection{
+  
  public:
-
+  
   SlowControlCollection();
   ~SlowControlCollection();
-
+  
   bool Init(zmq::context_t* context, int port=555, bool new_service=true);
   bool ListenForData(int poll_length=0);
   bool InitThreadedReceiver(zmq::context_t* context, int port=555, int poll_length=100, bool new_service=true);
@@ -42,14 +44,14 @@ class SlowControlCollection{
   bool TriggerSend(std::string trigger);
   std::string Print();
   template<typename T> T GetValue(std::string name){
-
-    return SC_vars[name]->GetValue<T>();    
-
-  }
   
-
+  return SC_vars[name]->GetValue<T>();    
+  
+}
+  
+  
  private:
-
+  
   std::map<std::string, SlowControlElement*> SC_vars;
   std::map<std::string, std::function<void(const char*)> > m_trigger_functions;
   std::mutex m_trigger_functions_mutex;
@@ -58,9 +60,11 @@ class SlowControlCollection{
   zmq::context_t* m_context;
   zmq::socket_t* m_pub;
   SlowControlCollectionThread_args* args;
-
+  
   static void Thread(Thread_args* arg);
-
+  
 };
-
+  
+}
+  
 #endif
